Restructure src/fullcalendar
