+++ src/components/Editor/FreeBusy/FreeBusy.vue
@@ -72,10 +72,10 @@ import {
 } from 'vuex'
 import { getLocale } from '@nextcloud/l10n'
 import Modal from '@nextcloud/vue/dist/Components/Modal'
-import VTimezoneNamedTimezone from '../../../fullcalendar/vtimezoneNamedTimezoneImpl.js'
-import freeBusyEventSource from '../../../fullcalendar/freeBusyEventSource.js'
+import VTimezoneNamedTimezone from '../../../fullcalendar/timezones/vtimezoneNamedTimezoneImpl.js'
+import freeBusyEventSource from '../../../fullcalendar/eventSources/freeBusyEventSource.js'
 import { getColorForFBType } from '../../../utils/freebusy.js'
-import freeBusyFakeBlockingEventSource from '../../../fullcalendar/freeBusyFakeBlockingEventSource.js'
+import freeBusyFakeBlockingEventSource from '../../../fullcalendar/eventSources/freeBusyFakeBlockingEventSource.js'
 
 export default {
 	name: 'FreeBusy',
