+++ src/fullcalendar/interaction/eventAllow.js
