+++ src/fullcalendar/interaction/eventDrop.js
@@ -19,21 +19,31 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-import { getDurationValueFromFullCalendarDuration } from './duration'
-import { getObjectAtRecurrenceId } from '../utils/calendarObject.js'
+import { getDurationValueFromFullCalendarDuration } from '../duration'
+import getTimezoneManager from '../../services/timezoneDataProviderService'
+import logger from '../../utils/logger.js'
+import { getObjectAtRecurrenceId } from '../../utils/calendarObject.js'
 
 /**
- * Returns a function to resize an event
+ * Returns a function to drop an event at a different position
  *
- * @param {Object} store The Vuex Store
+ * @param {Object} store The Vuex store
+ * @param {Object} fcAPI The fullcalendar api
  * @returns {Function}
  */
-export default function(store) {
-	return async function({ event, startDelta, endDelta, revert }) {
-		const startDeltaDuration = getDurationValueFromFullCalendarDuration(startDelta)
-		const endDeltaDuration = getDurationValueFromFullCalendarDuration(endDelta)
+export default function(store, fcAPI) {
+	return async function({ event, delta, revert }) {
+		const deltaDuration = getDurationValueFromFullCalendarDuration(delta)
+		const defaultAllDayDuration = getDurationValueFromFullCalendarDuration(fcAPI.getOption('defaultAllDayEventDuration'))
+		const defaultTimedDuration = getDurationValueFromFullCalendarDuration(fcAPI.getOption('defaultTimedEventDuration'))
+		const timezoneId = fcAPI.getOption('timeZone')
+		let timezone = getTimezoneManager().getTimezoneForId(timezoneId)
+		if (!timezone) {
+			timezone = getTimezoneManager().getTimezoneForId('UTC')
+			logger.error(`EventDrop: Timezone ${timezoneId} not found, falling back to UTC.`)
+		}
 
-		if (!startDeltaDuration && !endDeltaDuration) {
+		if (!deltaDuration || !defaultAllDayDuration || !defaultTimedDuration) {
 			revert()
 			return
 		}
@@ -58,11 +68,16 @@ export default function(store) {
 			return
 		}
 
-		if (startDeltaDuration) {
-			eventComponent.addDurationToStart(startDeltaDuration)
-		}
-		if (endDeltaDuration) {
-			eventComponent.addDurationToEnd(endDeltaDuration)
+		try {
+			// shiftByDuration may throw exceptions in certain cases
+			eventComponent.shiftByDuration(deltaDuration, event.allDay, timezone, defaultAllDayDuration, defaultTimedDuration)
+		} catch (error) {
+			store.commit('resetCalendarObjectToDav', {
+				calendarObject,
+			})
+			console.debug(error)
+			revert()
+			return
 		}
 
 		if (eventComponent.canCreateRecurrenceExceptions()) {
