+++ src/fullcalendar/interaction/navLinkWeekClick.js
@@ -19,7 +19,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-import { getYYYYMMDDFromDate } from '../utils/date.js'
+import { getYYYYMMDDFromDate } from '../../utils/date.js'
 
 /**
  * Handles a click on a week-number in the calendar-grid
