+++ src/fullcalendar/interaction/select.js
