+++ src/fullcalendar/rendering/eventLimitText.js
