+++ src/fullcalendar/rendering/eventRender.js
