+++ src/fullcalendar/rendering/windowResize.js
