+++ tests/javascript/unit/fullcalendar/interaction/eventDrop.test.js
@@ -19,10 +19,10 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-import eventDrop from "../../../../src/fullcalendar/eventDrop.js";
-import { getDurationValueFromFullCalendarDuration} from "../../../../src/fullcalendar/duration.js";
-import getTimezoneManager from '../../../../src/services/timezoneDataProviderService.js'
-import {getObjectAtRecurrenceId} from "../../../../src/utils/calendarObject.js";
+import eventDrop from "../../../../../src/fullcalendar/interaction/eventDrop.js";
+import { getDurationValueFromFullCalendarDuration} from "../../../../../src/fullcalendar/duration.js";
+import getTimezoneManager from '../../../../../src/services/timezoneDataProviderService.js'
+import {getObjectAtRecurrenceId} from "../../../../../src/utils/calendarObject.js";
 
 jest.mock("../../../../src/fullcalendar/duration.js")
 jest.mock('../../../../src/services/timezoneDataProviderService.js')
