+++ tests/javascript/unit/fullcalendar/interaction/eventResize.test.js
@@ -19,10 +19,10 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-import eventResize from "../../../../src/fullcalendar/eventResize.js";
+import eventResize from "../../../../../src/fullcalendar/interaction/eventResize.js";
 
-import { getDurationValueFromFullCalendarDuration} from '../../../../src/fullcalendar/duration.js'
-import {getObjectAtRecurrenceId} from "../../../../src/utils/calendarObject.js";
+import { getDurationValueFromFullCalendarDuration} from '../../../../../src/fullcalendar/duration.js'
+import {getObjectAtRecurrenceId} from "../../../../../src/utils/calendarObject.js";
 jest.mock('../../../../src/fullcalendar/duration.js')
 jest.mock("../../../../src/utils/calendarObject.js")
