+++ tests/javascript/unit/fullcalendar/interaction/navLinkDayClick.test.js
@@ -19,8 +19,8 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-import navLinkDayClick from '../../../../src/fullcalendar/navLinkDayClick.js'
-import { getYYYYMMDDFromDate } from '../../../../src/utils/date.js'
+import navLinkDayClick from '../../../../../src/fullcalendar/interaction/navLinkDayClick.js'
+import { getYYYYMMDDFromDate } from '../../../../../src/utils/date.js'
 jest.mock('../../../../src/utils/date.js')
 
 describe('fullcalendar/eventClick test suite', () => {
