+++ tests/javascript/unit/fullcalendar/interaction/navLinkWeekClick.test.js
@@ -19,8 +19,8 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-import navLinkWeekClick from '../../../../src/fullcalendar/navLinkWeekClick.js'
-import { getYYYYMMDDFromDate } from '../../../../src/utils/date.js'
+import navLinkWeekClick from '../../../../../src/fullcalendar/interaction/navLinkWeekClick.js'
+import { getYYYYMMDDFromDate } from '../../../../../src/utils/date.js'
 jest.mock('../../../../src/utils/date.js')
 
 describe('fullcalendar/eventClick test suite', () => {
